requirements.txt
django==4.2.0
djangorestframework==3.14.0
django-filter==23.2
django-environ==0.10.0
django-extensions==4.1  
psycopg2-binary==2.9.6
pillow==9.5.0
drf-spectacular==0.26.2
celery==5.2.7
redis==4.5.4
openpyxl==3.1.2
requests==2.31.0
psycopg2-binary==2.9.6
openpyxl==3.1.2
pytest-django==4.11.1